requirements.txt
allure-pytest==2.9.45
pytest~=7.2.2
selenium~=4.11.2
requests~=2.28.1
pytest-rerunfailures==10.3
pytest-xdist==2.5.0